+++ tests_old/promising_function/sync/test_sync_functions.py
@@ -8,22 +8,6 @@ from promising import get_active_promise
 # ── Core: Sync Function Wrapping & Argument Forwarding ──────────
 
 
-async def test_calling_sync_promising_function_returns_promise() -> None:
-    """
-    Calling a decorated sync function returns a Promise;
-    awaiting it returns the expected value.
-    """
-
-    @promising.function(use_thread_pool=True)
-    def greet() -> str:
-        return "hello"
-
-    assert isinstance(greet, promising.PromisingFunction)
-    result = greet()
-    assert isinstance(result, promising.Promise)
-    assert await result == "hello"
-
-
 async def test_forwards_positional_args() -> None:
     """
     Positional args are correctly forwarded to the
