+++ tests_old/promising_function/test_promising_function.py
@@ -5,22 +5,6 @@ import promising
 # ── Core: Async Function Wrapping & Argument Forwarding ──────────
 
 
-async def test_calling_promising_function_returns_promise() -> None:
-    """
-    Calling a decorated function returns a Promise;
-    awaiting it returns the expected value.
-    """
-
-    @promising.function
-    async def greet() -> str:
-        return "hello"
-
-    assert isinstance(greet, promising.PromisingFunction)
-    result = greet()
-    assert isinstance(result, promising.Promise)
-    assert await result == "hello"
-
-
 async def test_forwards_positional_args() -> None:
     """
     Positional args are correctly forwarded to the
